% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_showcase.R
\name{create_showcase}
\alias{create_showcase}
\title{Extract examples from documentation and place in gitbook}
\usage{
create_showcase(
  r_dir_path = './R',
  gitbook_dir_path = './examples/showcase/'
)
}
\arguments{
\item{r_dir_path}{[character], length 1, dir path of R file}

\item{gitbook_dir_path}{[character], length 1, output directory}
}
\description{
Extract examples from documentation and place in gitbook
}
\examples{
#BEGIN: example
\dontrun{
# simplest example
create_showcase()
}
#END: example
}
