% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_developing.R
\name{format_debugging}
\alias{format_debugging}
\title{Reformat file generated by use_r.function to developing structure}
\usage{
format_debugging(
  file_path,
  dry_run = TRUE
)
}
\arguments{
\item{file_path}{[character], length 1, file path to reformat}

\item{dry_run}{[logical], length 1, if TRUE, print the reformatted file}
}
\value{

}
\description{
Reformat file generated by use_r.function to developing structure
}
