% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_developing.R
\name{reformat_developing}
\alias{reformat_developing}
\title{Reformat file generated by use_r.function to developing structure}
\usage{
reformat_developing(
  file_path,
  dry_run = TRUE
)
}
\arguments{
\item{file_path}{[character], length 1, file path to reformat}

\item{dry_run}{[logical], length 1, if TRUE, print the reformatted file}
}
\value{

}
\description{
Reformat file generated by use_r.function to developing structure
}
\examples{
#BEGIN: example
\dontrun{
# observe that reformat_developing moves examples next to description
use_r.function('myfun.R')
reformat_packaging('myfun.R',dry_run=FALSE)
reformat_developing('myfun.R',dry_run=TRUE)
}
#END: example
}
