% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_packaging.R
\name{reformat_packaging}
\alias{reformat_packaging}
\title{Reformat file generated by use_r.function to CRAN valid structure}
\usage{
reformat_packaging(
  file_path,
  dry_run = TRUE
)
}
\arguments{
\item{file_path}{[character], length 1, file path to reformat}

\item{dry_run}{[logical], length 1, if TRUE, print the reformatted file}
}
\value{

}
\description{
Reformat file generated by use_r.function to CRAN valid structure
}
\examples{
#BEGIN: example
\dontrun{
# observe that it moves examples next to description
use_r.function('myfun.R')
reformat_package('myfun.R')
}
#END: example
}
