% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_md.gitbook.R
\name{use_md.githbook}
\alias{use_md.githbook}
\title{Gitbook template}
\usage{
use_md.githbook(
  dir_path
)
}
\arguments{
\item{dir_path}{[character], directory path to place template}
}
\description{
Gitbook template
}
