% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r.function.R
\name{use_r.function}
\alias{use_r.function}
\title{Create file for function using packaging template}
\usage{
use_r.function(
  file_path,
  edit_file = TRUE
)
}
\arguments{
\item{file_path}{[character], length == 1, file path}

\item{edit_file}{[logical], length == 1, if TRUE will open file if rstudio is available}
}
\description{
Create file for function using packaging template
}
