% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmd.gitbook.R
\name{use_rmd.githbook}
\alias{use_rmd.githbook}
\title{Gitbook template}
\usage{
use_rmd.githbook(
  dir_path
)
}
\arguments{
\item{dir_path}{[character], directory path to place template}
}
\description{
Gitbook template
}
\examples{
#BEGIN: example
\dontrun{
# create the template
library(usethis)
library(writethat)
use_rmd.githbook('my_book')
# start editing a file
usethis::edit_file('./my_book/01-roadmap.Rmd')
}
#END: example
}
